object hierarchy refactor and beginning of writing support

+++ PDFParser.Parser/Objects/BooleanObject.cs
@@ -5,7 +5,7 @@ public class BooleanObject : DirectObject
     
     public bool Value { get; }
     
-    public BooleanObject(bool value, long offset, long length) : base(offset, length)
+    public BooleanObject(bool value, long offset, int length) : base(offset, length)
     {
         Value = value;
     }
