+++ PDFParser.Parser/Objects/NullObject.cs
@@ -4,7 +4,7 @@ namespace PDFParser.Parser.Objects;
 
 public class NullObject : DirectObject
 {
-    public NullObject(long offset, long length) : base(offset, length)
+    public NullObject(long offset, int length) : base(offset, length)
     {
         Debug.WriteLine("Null Object Created!");
     }
